[feat/#3] JWT 및 OAuth 관련 오류 상태 코드 및 메시지 정의
